LPS-165261 Add checking to avoid NPE returned by dir.listFiles()

+++ modules/util/portal-tools-service-builder/src/main/java/com/liferay/portal/tools/service/builder/ServiceBuilder.java
@@ -3363,6 +3363,10 @@ public class ServiceBuilder {
 
 		File dir = file.getParentFile();
 
+		if (!dir.exists() || !dir.isDirectory()) {
+			return;
+		}
+
 		for (File oldFile : dir.listFiles()) {
 			if (!Objects.equals(file.getName(), oldFile.getName())) {
 				oldFile.delete();
